fix stop redefining  LiteLLMBase
