+++ litellm/router_strategy/lowest_latency.py
@@ -12,6 +12,7 @@ from litellm import ModelResponse, token_counter, verbose_logger
 from litellm.caching.caching import DualCache
 from litellm.integrations.custom_logger import CustomLogger
 from litellm.litellm_core_utils.core_helpers import _get_parent_otel_span_from_kwargs
+from litellm.types.utils import LiteLLMBase
 
 if TYPE_CHECKING:
     from opentelemetry.trace import Span as _Span
@@ -19,19 +20,6 @@ if TYPE_CHECKING:
     Span = _Span
 else:
     Span = Any
-
-
-class LiteLLMBase(BaseModel):
-    """
-    Implements default functions, all pydantic objects should have.
-    """
-
-    def json(self, **kwargs):  # type: ignore
-        try:
-            return self.model_dump()  # noqa
-        except Exception:
-            # if using pydantic v1
-            return self.dict()
 
 
 class RoutingArgs(LiteLLMBase):
