+++ litellm/router_strategy/lowest_tpm_rpm_v2.py
@@ -14,6 +14,7 @@ from litellm.caching.caching import DualCache
 from litellm.integrations.custom_logger import CustomLogger
 from litellm.litellm_core_utils.core_helpers import _get_parent_otel_span_from_kwargs
 from litellm.types.router import RouterErrors
+from litellm.types.utils import LiteLLMBase
 from litellm.utils import get_utc_datetime, print_verbose
 
 if TYPE_CHECKING:
@@ -22,19 +23,6 @@ if TYPE_CHECKING:
     Span = _Span
 else:
     Span = Any
-
-
-class LiteLLMBase(BaseModel):
-    """
-    Implements default functions, all pydantic objects should have.
-    """
-
-    def json(self, **kwargs):  # type: ignore
-        try:
-            return self.model_dump()  # noqa
-        except Exception:
-            # if using pydantic v1
-            return self.dict()
 
 
 class RoutingArgs(LiteLLMBase):
